Test for clear all completed
